requirements.txt
streamlit~=1.22.0
Pillow~=10.1.0
streamlit-code-editor~=0.1.10